+++ pkg/model/trace_test.go
@@ -1,8 +1,9 @@
-package util
+package model
 
 import (
 	"bytes"
 	"fmt"
+	"hash/fnv"
 	"math/rand"
 	"testing"
 
@@ -11,6 +12,7 @@ import (
 	v1 "github.com/grafana/tempo/pkg/tempopb/trace/v1"
 	"github.com/grafana/tempo/pkg/util/test"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestCombine(t *testing.T) {
@@ -83,7 +85,7 @@ func TestCombine(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		actual, _, err := CombineTraces(tt.trace1, tt.trace2)
+		actual, _, err := CombineTraceBytes(tt.trace1, tt.trace2, "", "")
 		if len(tt.errString) > 0 {
 			assert.EqualError(t, err, tt.errString)
 		} else {
@@ -166,7 +168,7 @@ func BenchmarkCombineTraces(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		// nolint:errcheck
-		CombineTraces(b1, b2)
+		CombineTraceBytes(b1, b2, "", "")
 	}
 }
 
@@ -182,7 +184,7 @@ func BenchmarkCombineTracesIdentical(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		// nolint:errcheck
-		CombineTraces(b1, b2)
+		CombineTraceBytes(b1, b2, "", "")
 	}
 }
 
@@ -272,5 +274,27 @@ func TestSortTrace(t *testing.T) {
 		SortTrace(tt.input)
 
 		assert.Equal(t, tt.expected, tt.input)
+	}
+}
+
+func TestUnmarshal(t *testing.T) {
+	trace := test.MakeTrace(100, nil)
+	bytes, err := proto.Marshal(trace)
+	require.NoError(t, err)
+
+	actual, err := Unmarshal(bytes, CurrentEncoding)
+	require.NoError(t, err)
+
+	assert.True(t, proto.Equal(trace, actual))
+}
+
+func BenchmarkTokenForID(b *testing.B) {
+	h := fnv.New32()
+	id := []byte{0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08}
+	buffer := make([]byte, 4)
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_ = tokenForID(h, buffer, 0, id)
 	}
 }