+++ tempodb/wal/append_block_test.go
@@ -8,6 +8,10 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+func TestAppendBlockBadFile(t *testing.T) {
+
+}
+
 func TestFullFilename(t *testing.T) {
 	tests := []struct {
 		name     string
@@ -17,7 +21,7 @@ func TestFullFilename(t *testing.T) {
 		{
 			name: "legacy",
 			b: &AppendBlock{
-				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v0", backend.EncNone),
+				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v0", backend.EncNone, ""),
 				filepath: "/blerg",
 			},
 			expected: "/blerg/123e4567-e89b-12d3-a456-426614174000:foo",
@@ -25,7 +29,7 @@ func TestFullFilename(t *testing.T) {
 		{
 			name: "ez-mode",
 			b: &AppendBlock{
-				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v1", backend.EncNone),
+				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v1", backend.EncNone, ""),
 				filepath: "/blerg",
 			},
 			expected: "/blerg/123e4567-e89b-12d3-a456-426614174000:foo:v1:none",
@@ -33,7 +37,7 @@ func TestFullFilename(t *testing.T) {
 		{
 			name: "nopath",
 			b: &AppendBlock{
-				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v1", backend.EncNone),
+				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v1", backend.EncNone, ""),
 				filepath: "",
 			},
 			expected: "123e4567-e89b-12d3-a456-426614174000:foo:v1:none",
@@ -41,7 +45,7 @@ func TestFullFilename(t *testing.T) {
 		{
 			name: "gzip",
 			b: &AppendBlock{
-				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncGZIP),
+				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncGZIP, ""),
 				filepath: "",
 			},
 			expected: "123e4567-e89b-12d3-a456-426614174000:foo:v2:gzip",
@@ -49,7 +53,7 @@ func TestFullFilename(t *testing.T) {
 		{
 			name: "lz41M",
 			b: &AppendBlock{
-				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncLZ4_1M),
+				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncLZ4_1M, ""),
 				filepath: "",
 			},
 			expected: "123e4567-e89b-12d3-a456-426614174000:foo:v2:lz4-1M",
@@ -57,7 +61,7 @@ func TestFullFilename(t *testing.T) {
 		{
 			name: "lz4256k",
 			b: &AppendBlock{
-				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncLZ4_256k),
+				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncLZ4_256k, ""),
 				filepath: "",
 			},
 			expected: "123e4567-e89b-12d3-a456-426614174000:foo:v2:lz4-256k",
@@ -65,7 +69,7 @@ func TestFullFilename(t *testing.T) {
 		{
 			name: "lz4M",
 			b: &AppendBlock{
-				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncLZ4_4M),
+				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncLZ4_4M, ""),
 				filepath: "",
 			},
 			expected: "123e4567-e89b-12d3-a456-426614174000:foo:v2:lz4",
@@ -73,7 +77,7 @@ func TestFullFilename(t *testing.T) {
 		{
 			name: "lz64k",
 			b: &AppendBlock{
-				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncLZ4_64k),
+				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncLZ4_64k, ""),
 				filepath: "",
 			},
 			expected: "123e4567-e89b-12d3-a456-426614174000:foo:v2:lz4-64k",
@@ -81,7 +85,7 @@ func TestFullFilename(t *testing.T) {
 		{
 			name: "snappy",
 			b: &AppendBlock{
-				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncSnappy),
+				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncSnappy, ""),
 				filepath: "",
 			},
 			expected: "123e4567-e89b-12d3-a456-426614174000:foo:v2:snappy",
@@ -89,10 +93,18 @@ func TestFullFilename(t *testing.T) {
 		{
 			name: "zstd",
 			b: &AppendBlock{
-				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncZstd),
+				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v2", backend.EncZstd, ""),
 				filepath: "",
 			},
 			expected: "123e4567-e89b-12d3-a456-426614174000:foo:v2:zstd",
+		},
+		{
+			name: "data encoding",
+			b: &AppendBlock{
+				meta:     backend.NewBlockMeta("foo", uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"), "v1", backend.EncNone, "dataencoding"),
+				filepath: "/blerg",
+			},
+			expected: "/blerg/123e4567-e89b-12d3-a456-426614174000:foo:v1:none:dataencoding",
 		},
 	}
 
@@ -106,13 +118,14 @@ func TestFullFilename(t *testing.T) {
 
 func TestParseFilename(t *testing.T) {
 	tests := []struct {
-		name             string
-		filename         string
-		expectUUID       uuid.UUID
-		expectTenant     string
-		expectedVersion  string
-		expectedEncoding backend.Encoding
-		expectError      bool
+		name                 string
+		filename             string
+		expectUUID           uuid.UUID
+		expectTenant         string
+		expectedVersion      string
+		expectedEncoding     backend.Encoding
+		expectedDataEncoding string
+		expectError          bool
 	}{
 		{
 			name:             "ez-mode",
@@ -129,6 +142,15 @@ func TestParseFilename(t *testing.T) {
 			expectTenant:     "foo",
 			expectedVersion:  "v1",
 			expectedEncoding: backend.EncSnappy,
+		},
+		{
+			name:                 "version, encoding and dataencoding",
+			filename:             "123e4567-e89b-12d3-a456-426614174000:foo:v1:snappy:dataencoding",
+			expectUUID:           uuid.MustParse("123e4567-e89b-12d3-a456-426614174000"),
+			expectTenant:         "foo",
+			expectedVersion:      "v1",
+			expectedEncoding:     backend.EncSnappy,
+			expectedDataEncoding: "dataencoding",
 		},
 		{
 			name:        "path fails",
@@ -161,8 +183,8 @@ func TestParseFilename(t *testing.T) {
 			expectError: true,
 		},
 		{
-			name:        "wrong splits - 5",
-			filename:    "123e4567-e89b-12d3-a456-426614174000:test:test:test:test",
+			name:        "wrong splits - 6",
+			filename:    "123e4567-e89b-12d3-a456-426614174000:test:test:test:test:test",
 			expectError: true,
 		},
 		{
@@ -184,7 +206,7 @@ func TestParseFilename(t *testing.T) {
 
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			actualUUID, actualTenant, actualVersion, actualEncoding, err := parseFilename(tc.filename)
+			actualUUID, actualTenant, actualVersion, actualEncoding, actualDataEncoding, err := parseFilename(tc.filename)
 
 			if tc.expectError {
 				assert.Error(t, err)
@@ -196,6 +218,7 @@ func TestParseFilename(t *testing.T) {
 			assert.Equal(t, tc.expectTenant, actualTenant)
 			assert.Equal(t, tc.expectedEncoding, actualEncoding)
 			assert.Equal(t, tc.expectedVersion, actualVersion)
+			assert.Equal(t, tc.expectedDataEncoding, actualDataEncoding)
 		})
 	}
 }